+++ src/main/java/ai/mcpdirect/backend/dao/entity/aitool/AIPortVirtualTool.java
@@ -1,0 +1,9 @@
+package ai.mcpdirect.backend.dao.entity.aitool;
+
+public class AIPortVirtualTool extends AIPortTool{
+    public long toolId;
+    public String tags;
+
+    public AIPortVirtualTool() {}
+
+}